+++ Sources/SwiftUIKit/_Deprecated/UserDefaultsPersisted.swift
@@ -1,0 +1,2 @@
+@available(*, deprecated, renamed: "Persisted")
+public typealias UserDefaultsPersisted = Persisted